Cleaned up code
Turrets no longer build - needs fixing

+++ TrainRaid/src/com/derby/trainraid/BoundingBox.java
@@ -12,7 +12,7 @@ public class BoundingBox {
 	
 	public BoundingBox(Entity entity)
 	{
-		_boundingBox = new Rect(entity.GetX(), entity.GetY(), entity.GetX() + entity.getWidth(), entity.GetY() + entity.getHeight());
+		_boundingBox = new Rect((int)entity.GetX(), (int)entity.GetY(), (int)entity.GetX() + entity.getWidth(), (int)entity.GetY() + entity.getHeight());
 	}
 	
 	public Rect getBoundingBox()
@@ -20,14 +20,14 @@ public class BoundingBox {
 		return _boundingBox;
 	}
 	
-	public void setBoundingBoxPos(int x, int y, int height, int width)
+	public void setBoundingBoxPos(float x, float y, int height, int width)
 	{
-		_boundingBox.set(x, y, x + width, y + height);
+		_boundingBox.set((int)x, (int)y, (int)x + width, (int)y + height);
 	}
 	
 	public void setBoundingBox(Entity entity, int width, int height)
 	{
-		_boundingBox = new Rect (entity.GetX(), entity.GetY(), entity.GetX() + width, entity.GetY() + height);
+		_boundingBox = new Rect ((int)entity.GetX(), (int)entity.GetY(), (int)entity.GetX() + width, (int)entity.GetY() + height);
 	}
 	
 }
